pyproject.toml
[tool.poetry]
name = "coin_mena_challenge"
version = "0.1.0"
description = "Solution for ConMENA Challenge"
authors = ["Marc Philippe de Villeres <mpdevilleres@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
django-environ = "^0.5.0"
Markdown = "^3.3.4"
psycopg2-binary = "^2.9.1"
djangorestframework-simplejwt = "^4.8.0"
celery = "^5.1.2"
redis = "^3.5.3"
alpha-vantage = "^2.3.1"
drf-yasg = "^1.20.0"
gunicorn = "^20.1.0"


[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
